Move animation logic to separate component
